pyproject.toml
[project]
name = "chronosense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.2.1",
    "ipykernel>=7.0.1",
    "keras>=3.12.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-groq>=1.0.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pyarrow>=22.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.50.0",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "torch>=2.9.0",
    "yfinance>=0.2.66",
]
